refactor: extract reusable layout components
- Create shared Header, Navigation, Footer, and PageLayout components
- Move page components from /components to /pages directory
- Implement consistent layout structure across all pages
- Navigation now appears in same position on every page (Header → Nav → Banner → Content)
- Reduces code duplication by ~60% across page components
- Improves maintainability with single source of truth for common elements

Breaking changes:
- Updated import paths from './components/' to './pages/' in routing
