+++ src/client/components/Footer.jsx
@@ -1,0 +1,37 @@
+import React from 'react';
+import { Link } from 'react-router-dom';
+
+const Footer = () => {
+  return (
+    <>
+      {/* Social Links */}
+      <div className="social-links">
+        <a href="https://linkedin.com/in/matthew-bixby/" target="_blank" rel="noopener noreferrer" className="social-icon">
+          <div className="pixel-icon">in</div>
+        </a>
+        <a href="https://github.com/mattbixby123" target="_blank" rel="noopener noreferrer" className="social-icon">
+          <div className="pixel-icon">
+            <svg viewBox="0 0 24 24" width="20" height="20">
+              <path fill="currentColor" d="M12 2C6.477 2 2 6.477 2 12c0 4.42 2.865 8.166 6.839 9.489.5.092.682-.217.682-.482 0-.237-.008-.866-.013-1.7-2.782.603-3.369-1.342-3.369-1.342-.454-1.155-1.11-1.462-1.11-1.462-.908-.62.069-.608.069-.608 1.003.07 1.531 1.03 1.531 1.03.892 1.529 2.341 1.087 2.91.832.092-.647.35-1.088.636-1.338-2.22-.253-4.555-1.11-4.555-4.943 0-1.091.39-1.984 1.029-2.683-.103-.253-.446-1.27.098-2.647 0 0 .84-.268 2.75 1.026A9.578 9.578 0 0112 6.836c.85.004 1.705.114 2.504.336 1.909-1.294 2.747-1.026 2.747-1.026.546 1.377.202 2.394.1 2.647.64.699 1.028 1.592 1.028 2.683 0 3.841-2.337 4.687-4.565 4.935.359.309.678.919.678 1.852 0 1.336-.012 2.415-.012 2.743 0 .267.18.578.688.48C19.138 20.161 22 16.416 22 12c0-5.523-4.477-10-10-10z"></path>
+            </svg>
+          </div>
+        </a>
+        <Link to="/contact" className="social-icon">
+          <div className="pixel-icon">@</div>
+        </Link>
+      </div>
+
+      {/* Footer */}
+      <div className="retro-footer">
+        © {new Date().getFullYear()} Matthew Bixby 💻
+      </div>
+
+      {/* Watermark */}
+      <div className="watermark">
+        All rights reserved.
+      </div>
+    </>
+  );
+};
+
+export default Footer;
